python/requirements.txt
# VALORANT Scrim Analyzer - Python Dependencies

# Core
anthropic>=0.40.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0

# Data analysis
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.12.0

# OCR (optional - for local OCR without Claude Vision)
pytesseract>=0.3.10
Pillow>=10.2.0

# Video processing
opencv-python>=4.9.0.80

# API server (if running Python as separate service)
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# Utilities
tqdm>=4.66.0
